backend/job: Avoid race condition in kill()
Because kill() does not block when sending the kill request, the request
will be silently ignored if the watch() goroutine is not ready yet. To
avoid this, we use a buffered channel. However, this channel has to be
created before the Execute() call. Hence, we introduce the NewJob()
function to initialize a job.