VM: change wait() to waitpid()
The wait() function is a specialization of waitpid that will wait until
"status information for one of the terminated child process of the calling process is
available" (extract from the "man 3 wait" page from POSIX Programmer's manual).

This behaviour is ok if we only have one process but if we have multiples processes this could
create unwanted behaviours (script closing before finishing because another process just finish).

We will thus use the waitpid that will wait for the process that we just forked (i.e. the script
specified in the control).